Create ReadyBackgroundAdDisplay class
For displaying background ad logic
